refactor: 로또 번호 검증값 상수 활용
