+++ src/main/java/step2/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step2;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,10 +6,10 @@ import java.util.List;
 
 import static java.util.Collections.*;
 
-public class LottoNumberMachine {
+public class LottoNumberGenerator {
     private static final List<Integer> LOTTO_NUMBERS = new ArrayList<>();
-    private static final int MIN_NUMBER = 1;
-    private static final int MAX_NUMBER = 45;
+    public static final int MIN_NUMBER = 1;
+    public static final int MAX_NUMBER = 45;
     private static final int LOTTO_LENGTH = 6;
 
     static {
@@ -18,7 +18,7 @@ public class LottoNumberMachine {
         }
     }
 
-    public static List<Integer> provideLottoNumber() {
+    public static List<Integer> provideLottoNumbers() {
         Collections.shuffle(LOTTO_NUMBERS); // 셔플
         List<Integer> numbers = new ArrayList<>(LOTTO_NUMBERS.subList(0, LOTTO_LENGTH)); // 로또 자동
         sort(numbers); // 오름차순
