+++ src/main/java/step2/domain/lotto/LottoNumber.java
@@ -1,11 +1,13 @@
 package step2.domain.lotto;
 
+import static step2.LottoNumberGenerator.*;
+
 public class LottoNumber {
     private Integer number;
 
     public LottoNumber(Integer number) {
-        if (number < 1 || number > 45) {
-            throw new IllegalArgumentException("1이상 45이하의 숫자만 가능합니다");
+        if (number < MIN_NUMBER || number > MAX_NUMBER) {
+            throw new IllegalArgumentException(MIN_NUMBER + "이상 " + MAX_NUMBER + "이하의 숫자만 가능합니다");
         }
         this.number = number;
     }
